+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -3,9 +3,9 @@ CHANGELOG
 
 2.6.0
 -----
-
  * added "html5" option to Date, Time and DateTimeFormType to be able to
    enable/disable HTML5 input date when widget option is "single_text"
+ * added the html5 "range" FormType
 
 2.5.0
 ------