+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -7,6 +7,8 @@ CHANGELOG
  * added "html5" option to Date, Time and DateTimeFormType to be able to
    enable/disable HTML5 input date when widget option is "single_text"
  * added "label_format" option with possible placeholders "%name%" and "%id%"
+ * deprecated the "cascade_validation" option in favor of setting "constraints"
+   with the Valid constraint
 
 2.5.0
 ------